pyproject.toml
[tool.poetry]
name = "ndastro-ui"
version = "0.1.0"
description = "An astrology tool from DhuruvahApps"
authors = ["Jaganathan, Bantheswaran <jaganathan.eswaran@gmail.com>"]
readme = "README.md"
packages = [
    { include = "ndastro" }
]

[tool.poetry.dependencies]
python = ">=3.10.0,<=3.13.1"
pyside6 = "6.8.1"
skyfield = "^1.49"
python-i18n = "^0.3.9"
pytz = "^2024.2"
qdarkstyle = "^3.2.3"
darkdetect = "^0.8.0"
qt-material-icons = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
py-hot-reload = "^1.0.6"
pytest-mock = "^3.14.0"
typing-extensions = "^4.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 150

[tool.ruff.lint]
select = ["ALL"]
ignore=["INP001"]

[tool.ruff.lint.pep8-naming]
extend-ignore-names = ["enterEvent", "leaveEvent", "hoverEnterEvent", "hoverLeaveEvent", "resizeEvent"]

[tool.ruff.lint.per-file-ignores]
"**/{tests}/*" = ["S101", "D103", "D100", "ANN201"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 72

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests"
]
